remove providers
