+++ sdk/python/feast/errors.py
@@ -119,21 +119,9 @@ class FeastProviderNotImplementedError(Exception):
         super().__init__(f"Provider '{provider_name}' is not implemented")
 
 
-class FeastProviderNotSetError(Exception):
-    def __init__(self):
-        super().__init__("Provider is not set, but is required")
-
-
 class FeastRegistryNotSetError(Exception):
     def __init__(self):
         super().__init__("Registry is not set, but is required")
-
-
-class FeastFeatureServerTypeSetError(Exception):
-    def __init__(self, feature_server_type: str):
-        super().__init__(
-            f"Feature server type was set to {feature_server_type}, but the type should be determined by the provider"
-        )
 
 
 class FeastFeatureServerTypeInvalidError(Exception):
