--- sdk/python/feast/infra/aws.py
@@ -1,9 +1,0 @@
-from feast.infra.passthrough_provider import PassthroughProvider
-
-
-class AwsProvider(PassthroughProvider):
-    """
-    This class only exists for backwards compatibility.
-    """
-
-    pass
