--- sdk/python/feast/infra/gcp.py
@@ -1,9 +1,0 @@
-from feast.infra.passthrough_provider import PassthroughProvider
-
-
-class GcpProvider(PassthroughProvider):
-    """
-    This class only exists for backwards compatibility.
-    """
-
-    pass
